Add addl. COMP logic and fix recursion bug
The COMPUTER will now analyze the board for the move that
has the highest probability of winning the game, resulting in a
more humanlike feel.

The target probability is set at 15% and the analysis only
activates between turns 5 and 9 as a way to limit the points of
recursion used. There's probably a better algorithm, and I can
revisit the limits later, but for now, the COMP can resemble a
human player to the point of not being overly repetitive.

TODO: Refactor - this is a mess of functions and can be better
organized as object methods.

TODO: Now that the underlying logic is mostly complete, we can
begin styling the page.