Remove COMP win probability checker
Although the function could aid in realism, it was
constructed in a way that sometimes caused stack overflow due to
the recursive function used by COMP in choosing its next move. It
would also prevent the game from declaring a draw, as the final
move would never meet the win probability threshold, and anything
lower would be useless - meaning the COMP would never make the
final move and the game would stop short of ending.

Reverting back to the basic computer logic until a better
implementation can be found.